Add Rust Generator example
`Generator` is an unstable feature available in nightly Rust, which is being
used as the basis for the `async`/`await` features that are in active
development.

It can be used to provide a much more readable version of the range/iterator
example.  To do so, we need to create a wrapper type that implements
`Iterator` around a `Generator`.  I believe that my implementation is unsafe,
so should only be used for example purposes for now; it would need to pin the
generator to ensure it doesn't move during iteration to be safe.  It's
sufficient for the purpose of this example, and there have been changes to the
`Generator` interface that should be able to make this safe, they just haven't
been implemented yet.